Implement CH4 Calcs module.
This module includes unit tests and the app_test.sh integration test.
It does not include a System Test in test_excel_integration.py.
We've been using SolarPVUtil for the system test, and SolarPVUtil
sets CH4 to be CO2 equivalent. The CH4 Calcs tab in SolarPVUtil is
hidden and contains only #VALUE errors in Excel, which is not
useful for a test.

When we have more Project Drawdown Solutions in the pile to work on,
we could add a system test for CH4 Calcs using one of those.