Add data_sources Baseline/Ambitious/Conservative.
The TAM And Adoption Data modules take data from multiple
sources,categorized as Baseline, Ambitious, and Conservative.
There is overlap in the data sources used by different Drawdown
solutions, sufficient to we centralize the categorization
of data sources into categories.

This commit adds the data_sources implementation and unit tests,
but does not use it in the implementation of the model yet.