Cache operatingcost annual_breakout computation.
Computing the annual_breakout for the soln and conv cases
is quite expensive, and was being done a number of times
due to the number of other methods which depend on the result.
Compute the breakout once in __init__ and cache it.

cProfile.run('solution.solarpvutil.SolarPVUtil().to_dict()')
shows improvement from 20 seconds to 15 on a MacBook Air 2014.

Note that functools.lru_cache didn't work for this because
the arguments are dataframes and mutable, therefore unhashable.
We're instead taking advantage of the fact that all args are passed
to __init__ and therefore can not change.