Implement result caching.
Profiling shows that the TAM pds_tam_per_region and
ref_tam_per_region take up almost half of the time to
render the Jupyter notebook in its current form.
That can be greatly sped up by caching, as each is
called twice.

When we previously tried using functools.lru_cache it
wasn't feasible, as the arguments to many functions in
the models weremutable dataframes which could not be
hashed to make a cache key. Since then we're restructured
to pass in data to __init__ and make each method
take no arguments.

Before this CL:
  JSON fetch /solarpvutil: 14.387 seconds
  Jupyter notebook generation: 32.692

After this CL:
  JSON fetch /solarpvutil: 9.210 seconds
  Jupyter notebook generation: 18.894 seconds