solution_xls_extract: all modules & add tests.
Generate mostly useable Python code for a solution
and all of its scenarios. Only really works for RRS
solutions at this point, we need to build more of the
infrastructure for Land and Food and others.

Add pds_source_post_2014, ref_source_post_2014, and
source_until_2014 fields to Advanced Controls. They are
set by the scenario, but we've been handling them as
special cases. It works better to include them in
Advanced Controls.

Add an integration test in solution_xls_extract_test.sh
to check that solution_xls_extract is working.