Move energy-related CSV files to data/energy.
The vma_*.csv files don't need to be in the solution
directory, we have a very fine data/energy directory
where they can go. I don't really know why I put them
there in the first place.

Ideally solution/* will only hold a directory for each
solution, plus shared implementation like rrs.py and a
possible future land.py and food.py.

It is mostly the conventional VMA csv files which are
moving to data/energy. The solution-related VMAs tend
to be specific to that one solution, and are in
solution/<solution_name>. One exception is
vma_solar_util_vs_roof.csv which is used by both
SolarPVUtil and SolarPVRoof. It is moving to data/energy
with a filename to refer to its use.