Implement regional adoption_data.
The first several solutions we implemented only had adoption_data for
the 'World' region. As we're Faulting the Excel Model in On Demand (TM),
we left the rest unimplemented until we had a solution which needed
them. The concentratedsolar solution being worked on now does have
regional adoption data.

This involved updating a number of Adoption Data CSV files to add
the regional header. If we don't have any data we leave those columns
out, and Pandas will fill in N/A.

Also, while here:
+ rename the unit test CSV files to include "based on," as they
  are the distilled results and not the original data sets.
+ fix two issues in solution_xls_extract: Adoption Data extraction
  was missing three years of data, and the spreadsheet doesn't
  actually use the Advanced Controls growth setting for TAM
  only for adoption.