Jupyter: Altair charts, treemap, sorted solutions.
1. Altair charts
After investigating a number of alternatives to matplotlib in
https://codingrelic.geekhold.com/2019/02/line-graphs-in-jupyterlab.html
settle on Altair charts for interactive use. matplotlib's static
charts are great, but we get more compelling interactive use from
Altair. Remove matplotlib charts and replace with Altair.

We still use bqplot for the nested donut chart on the Overview tab,
as Altair doesn't support a chart like that.

2. Treemap
Add a treemap on the Overview tab, where the first level is the
sector and the second is the solution. The implementation generates
a Vega description (ex: https://vega.github.io/vega/examples/treemap/)
and passes it to Jupyter's display() to render. Several alternatives
were investigated first:
+ squarify (https://github.com/laserson/squarify), but this does only
  a single level treemap. We wanted sectors grouped first, then
  solutions within the sector.
+ beakerx, which did generate a two level treemap but did not allow
  control of the colors (it would select colors from a random pool)
  nor control of the labels (it labeled leaf nodes, and only some of
  them. We wanted the sectors labeled).

3. Sorting of solutions
Based on feedback, sort the list of solutions on the Overview tab
first by the total CO2 reduction of the sector, then by CO2 reduction
of solutions within the sector.