helpertables: implement PDS/REF overrides.
Continuing th practice of faulting in implementation
on demand, implement support for overriding PDS adoption
predictions from REF and REF from PDS for a selectable set
of years. In the spreadsheet this is implemented in
a concise description in the ScenarioRecord, which gets
expanded in SolarPVUtil 'Advanced Controls'!F292:F338.

This was noticed while working on ConcentratedSolar, which
uses this feature.

In the code generator, extract the concise definition from
the ScenarioRecord to store in advanced_controls.py.
helpertables.py operates directly on the concise definition.

While there, add notations to the comments about Excel cell
locations that come from SolarPVUtil. We are working on
a number of solutions in parallel, need to be specific
which one is being referenced.