adoptiondata: handle source = None.
For scenarios which set adoption_basis = 'DEFAULT Linear',
the adoption source and trend and growth are often left
unset and therefore None. Most of the adoption trend calculation
is not used for 'DEFAULT Linear' adoption we just need it
to do something rational like return NaN, not die and kill
the model.

Noticed while working on ImprovedCookStoves, which has one scenario
using DEFAULT Linear.