tam: regional data source definitions.
We've been handling regional data sources via columns for each
region in the CSV file, with a single dict defining the groups for
data sources like 'Ambitious Cases' or 'Conservative Cases'
This has worked to this point, but ImprovedCookStoves classifies
the same data source as Amitious for one region and Conservative
for another region.

As most solutions looked at so far do not do this, to allow a
succinct definition of sources we're allowing a region which
does not have a specific set of sources to rely on the single
global set of source definitions. Regions can override the
global sources by adding a 'Region: <name>" entry in the
data_sources.

Add a test case where two regions supply their own source definitions
in a way which will fail if the global sources are used (and verified
that it does fail if the global sources are used).

While here:
+ specify SolarPVUtil as the source of data in comments of
  cell locations.