adoptiondata: regional data source definitions.
This matches an earlier commit adding the same functionality
to model/tam.py.

We've been handling regional data sources via columns for each
region in the CSV file, with a single dict defining the groups for
data sources like 'Ambitious Cases' or 'Conservative Cases'
There are solutions where this is not sufficient, such as
ImprovedCookStoves for TAM Data.

To allow a succinct definition of sources we're allowing a region
which does not have a specific set of sources to rely on the single
global set of source definitions. Regions can override the global
sources by adding a 'Region: <name>" entry in the data_sources.

Add a test case to verify the functionality.