tam: PDS uses 'World' forecast for 2012-2014.
The PDS Mean computation is supposed to use the World
data for 2012-2014, to match the Excel implementation.
This has not been noticed thus far as all of the existing
solutions to date have featured curated datapoints for
2012-2014 which were identical between PDS and World.
SolarHotWater is the first solution where the two differed.

Add a test case to demonstrate the problem and verify
the fix.