TAM: remove caching metaclass
If the TAM uses the caching metaclass, and "make test" is used
to run all of the test_excel_integration cases, then SmartThermostats
fails in the TAM Data tests. Something, somewhere appears to be
modifying the cached TAM object.

SmartThermostats run by itself will pass.

I'm convinced that we need some kind of cache, as computing the TAM
is one of the most expensive parts of RRS solutions and performance
is a problem we need to solve, but it needs to be more robust
before re-introducing it.