Add bioplastic.
As with the earlier ConcentratedSolar, the Custom PDS Adoption
data has a dependency loop such that it varies depending on
the scenario. The CSV files in this CL correspond to the first
scenario. We'll resolve the others when we settle on a
fix which resolves all solutions like this one.