unitadoption: new_iunits_reqd accumulation.
Accumulating replacement units in a separate dataframe
doesn't handle one case correctly: there is a conditional
to decide whether the prior year had enough units to be
replaced, and the conditional needs to take previous
replacements into account.

Remove the separate dataframe and accumulate replacements
into the result DF as we go along.

Add a test case with data from LED Lighting (Commercial)
which has several years in the Middle East and Africa region
with 0.0 funits before replacements, and therefore the
conditional check would fail when it should have succeeded.
(Whether it is accurate to have years with 0.0 funits is a
different question).