operatingcost: break up lifetime_cost_forecast
lifetime_cost_forecast was written very early in the project,
when we thought Excel was going to be the first frontend and UI.
It combines four separate (though related) results into one
DataFrame:
  + Marginal First Cost
  + Marginal Operating Cost Savings
  + Net Cash Flow
  + Net Present Value

and gives the columns human-readable names appropriate to be
written to 'Operating Cost'!A126:E250.

The strategy changed as we went along, we no plan to use
Excel as a frontend and have removed the JSON support which
would have been used to do so.

Break lifetime_cost_forecast into separate routines for each
result, break the unit test up into separate tests, and change
test_excel_integration to use the separate routines.