tam: single source with 0.0 is fine.
In commit 16af3feafe1856f93c3ea336fb0a77d3c47fe97d
("tam: don't include 0.0 in the mean.") we implemented the
Excel formula for Mean computation:
  SUM($C521:$Q521)/COUNTIF($C521:$Q521,">0")

However it is slightly more complicated than this. The COUNTIF
code is only used when processing a group like 'ALL SOURCES' or
'Conservative Cases' and not when an individual source is chosen.
For an individual source, 0.0 is retained and not discarded.

Noticed while working on Water Efficiency, where the India region
has only one source and it predicts 0.0 for 2014 - 2021.
Note that the researcher appears to have had trouble with this
as well: where all of the other regions also have one source and
are set to use 'ALL SOURCES', the India region is set to the name
of that one specific source to work around the COUNTIFS(">0")
behavior.

We're still slightly different than Excel:
+ if you choose ALL SOURCES, and there is only one source, and it
  is 0.0, Excel will get a Medium result of NaN. Python will treat
  it as a single source and get a Medium result of 0.0.

As this appears to be the behavior the researcher for Water Efficiency
actually wanted, we're going to try it and see if it leads to issues
with any of the remaining solutions.

Also include the test case added in interpolation using the Water
Efficiency data, while trying to figure out the root cause behind
the test_excel_integration failure which was eventually traced to
this.