graphviz: graphical overview of the model.
Add an SVG diagram of how the modules in the model connect.
The initial idea was to dynamically generate the diagram
based on what was used in the solution, but it worked out
better to generate a fully-moulated diagram and then disable
the pieces which are not used in each given solution.

There are two top level SVG diagrams, one for Land and one for
RRS. A combined diagram was generated but wasn't very nice
for either purpose, having custom layouts for each worked better.

Add a tab in JupyterLab showing the diagram full size, and add
small diagrams on each module page highlighting the module being
looked at and where it fits in to the overall model.