modelhealth: include regional data survey
Show a histogram of scenarios and how close the sum of their major
regions is to their World region. There are spikes at 0% (i.e.
no regional data) and between 90-120% (reasonably good regional data)
with some weird outliers.

I'd expected to run tools/health/survey.py in gitlab CI right away, using
https://about.gitlab.com/2017/11/02/automating-boring-git-operations-gitlab-ci/
and https://gist.github.com/plajjan/42592665afd5ae045ee36220e19919aa#file-open-mr-sh
as a guide for how to make the bot create a CL and send a merge request.

However it looks like we're weeks away from a much easier way for
bots to create merge requests by pushing to a git branch:
https://gitlab.com/gitlab-org/gitlab-ce/issues/43263
The issue is being updated almost daily and code is being written
to implement it.

So we'll hold off a little while on setting up the bot, and run it
manually to be able to add the regional stats to the health dashboard.