test_solutions: parameterize from solution.factory.
Turn the giant set of tests into a single parameterized test,
generated from solution.factory.

A parameterized test is used so that if one fails, the rest of the
cases will be run to get a sense for how badly things have been broken.
If we used a loop inside a single test, it would stop at the first
failure.

Note that it is still possible to run the test for a single solution.
Instead of: pytest solution/test_solutions.py::test_airplanes
Use:        pytest solution/test_solutions.py::test_solutions[airplanes]

Parameterized tests have a blind spot: if we were to accidentally
delete most of the solutions, the test would still pass. Therefore,
include a test that we have a rational number of solutions. We
currently have 41 working solutions, the test asserts that we have
more than 35.