Support Bass Diffusion S-Curve in solutions.
Add support for Bass Diffusion, a second S-Curve adoption
module.

In Excel, the adoption basis is set to 'DEFAULT S-Curve' and then
a separate setting controls whether to use Logistic or Alternate.
That doesn't seem optimal. Instead, we'll set the adoption
basis to 'Logistic S-Curve' or 'Bass Diffusion S-Curve'
and drive the decisions from there.