let Excel_one_cent handle tropicaltreestaples
Due to floating point precision, sometimes subtracting what
appear to be identical values for unit adoption results not
in zero, but in 0.000000000000007105427357601000 which, when
multiplied by a large unit cost, can result in a First Cost
of (say) 2.5e-6 instead of the zero which might otherwise be
expected.

We've now run into this with two different solutions,
tropicaltreestaples and regenerativeagriculture. Make the test
case figure this out, by masking off comparisons in the two
First Cost tables for values less than one penny.

Verified that tropicaltreestaples passes now, without needing
to remove the First Cost comparisons.