ua: correct _direct_emissions_saved_land ghg_rplu
In the one-time + protection case, _direct_emissions_saved_land
was always multiplying by the co2eq factor not the ghg_rplu
passed into the routine. Noticed when we added CH4 support and
ForestProtection began failing test_excel_integration.

Correct the factor, and add a unit test for all of the GHG routines.