Regenerate solution/silvopasture from Public file
solution/silvopasture was the first Land solution, and was developed
by hand. The effort in silvopasture led to the code generator we use
now. Coming full circle, replace the manually written  code with
code generated.

Note that this also resolves an issue we're having with getting the
system to run on Windows: one of the ca_pds_data files contained
a ">" symbol in the filename, which is not allowed in a Windows
filename. We normalize the names now to remove characters which
can be problematic.