UI: Frizzle chart color scheme
The Frizzle strips are visually distinct, they don't need to
be in contrasting colors. Make them be the color of the Sector
instead, and make the text white so as to be visually similar
to the treemap and donut charts in the overview.

Also, fillna(0.0) as a few solutions (like Refrigerant Management)
have NaN in the TAM.