UI: remove fully-specified ui.charts module name.
Parts of a recent commit were developed in a code cell in
the notebook and then copied to the external file. Running
in the notebook requires a fully specified module name, but
we don't want to do so in the external file as that makes
the code brittle to changes in location or filename.