UI: remove bit.ly links and Doc link
We were using bit.ly links under the mistaken impression that
they could be redirected to a new target in the future. In reality,
changing the target is a premium feature, and is only allowed for
links on a custom domain not bit.ly itself. We have no reason to
use bit.ly now, just embed the direct links.

Remove the link to the Getting Started Guide, it documented how to
install the system to run locally. This did not work well for the
people who tried to do so, we're de-emphasizing local operation
and focussing on the cloud-hosted JupyterHub version.