health: survey params which vary between scenarios.
Count the number of solutions where each parameter in advanced_controls
varies between scenarios within the solution. We suspect that some
fraction of the time, this was unintentional and a result of the
Fellow editing one scenario without realizing that the ScenarioRecord
does not update that value in the other scenarios.