vma: match Excel weighted stddev.
Excel's implementation of weighted stddev isn't really correct, as it
includes the weights for sources which are excluded or discarded.
https://docs.google.com/document/d/19sq88J_PXY-y_EnqbSJDl0v9CdJArOdFLatNNUFhjEA/edit#heading=h.qkdzs364y2t2

We'd previously handled this by printing a warning message when
extracting a solution from an Excel file, that excluded or discarded
sources need to have their weights set to zero. This isn't really
practical when importing 70+ Excel solutions, as we don't have a good
way to determine whether the modification to this file has already been
made and it isn't reasonable to expect someone to manually examine every
VMA in every file every time the solutions are re-imported.

Additionally, Excel discards sources whose value is greater than 3
standard deviations away from the mean, but only when weights are not
used. If weights are enabled, outliers are not discarded.
We have documented this in
https://docs.google.com/document/d/19sq88J_PXY-y_EnqbSJDl0v9CdJArOdFLatNNUFhjEA/edit#heading=h.9w8cj2bcuegz

Update the implementation to match Excel's behavior, and add tests.