fc: first cost per iunit if iunits is zero.
In Excel, the first cost per iunit is DIV/0 if the number of iunits is
zero or NaN. See First Cost Handling of NaN data in:
https://docs.google.com/document/d/19sq88J_PXY-y_EnqbSJDl0v9CdJArOdFLatNNUFhjEA/edit#

Nuclear in its Optimum scenario has iunits reduce to zero, which led to
a failure in test_excel_integration.

Also reformat whitespace in the test file.