UI: ipywidget Layout align_items="stretch"
We'd been using justify_items="stretch", which did work but was
never correct. The flex layout only support stretch for
align_items.

Something changed in gitlab's continuous integration test environment
that this mistake now triggers an error. I'm not entirely sure what
changed, but it found a real error so we correct it here.